+++ Content.IntegrationTests/Tests/Materials/MaterialTests.cs
@@ -1,5 +1,4 @@
 #nullable enable
-using Content.Server.Stack;
 using Content.Shared.Stacks;
 using Content.Shared.Materials;
 using Robust.Shared.GameObjects;
@@ -12,7 +11,7 @@ namespace Content.IntegrationTests.Tests.Materials
     /// so we need some test coverage for them.
     /// </summary>
     [TestFixture]
-    [TestOf(typeof(StackSystem))]
+    [TestOf(typeof(SharedStackSystem))]
     [TestOf(typeof(MaterialPrototype))]
     public sealed class MaterialPrototypeSpawnsStackMaterialTest
     {
