+++ Content.Server/Construction/Completions/GivePrototype.cs
@@ -1,4 +1,3 @@
-using Content.Server.Stack;
 using Content.Shared.Construction;
 using Content.Shared.Hands.Components;
 using Content.Shared.Hands.EntitySystems;
@@ -26,7 +25,7 @@ public sealed partial class GivePrototype : IGraphAction
 
         if (EntityPrototypeHelpers.HasComponent<StackComponent>(Prototype))
         {
-            var stackSystem = entityManager.EntitySysManager.GetEntitySystem<StackSystem>();
+            var stackSystem = entityManager.EntitySysManager.GetEntitySystem<SharedStackSystem>();
             var stacks = stackSystem.SpawnMultiple(Prototype, Amount, userUid ?? uid);
 
             if (userUid is null || !entityManager.TryGetComponent(userUid, out HandsComponent? handsComp))
