+++ Content.Server/Construction/Completions/SetStackCount.cs
@@ -1,5 +1,5 @@
-﻿using Content.Server.Stack;
-using Content.Shared.Construction;
+﻿using Content.Shared.Construction;
+using Content.Shared.Stacks;
 using JetBrains.Annotations;
 
 namespace Content.Server.Construction.Completions
@@ -12,7 +12,7 @@ namespace Content.Server.Construction.Completions
 
         public void PerformAction(EntityUid uid, EntityUid? userUid, IEntityManager entityManager)
         {
-            entityManager.EntitySysManager.GetEntitySystem<StackSystem>().SetCount(uid, Amount);
+            entityManager.EntitySysManager.GetEntitySystem<SharedStackSystem>().SetCount(uid, Amount);
         }
     }
 }
