+++ Content.Server/Construction/ConstructionSystem.cs
@@ -1,7 +1,7 @@
 using Content.Server.Construction.Components;
-using Content.Server.Stack;
 using Content.Shared.Construction;
 using Content.Shared.DoAfter;
+using Content.Shared.Stacks;
 using JetBrains.Annotations;
 using Robust.Server.Containers;
 using Robust.Shared.Random;
@@ -18,7 +18,7 @@ namespace Content.Server.Construction
         [Dependency] private readonly IRobustRandom _robustRandom = default!;
         [Dependency] private readonly SharedDoAfterSystem _doAfterSystem = default!;
         [Dependency] private readonly ContainerSystem _container = default!;
-        [Dependency] private readonly StackSystem _stackSystem = default!;
+        [Dependency] private readonly SharedStackSystem _stackSystem = default!;
         [Dependency] private readonly SharedToolSystem _toolSystem = default!;
 
         public override void Initialize()
